+++ tasq/job.py
@@ -27,8 +27,12 @@ class Job:
     def __init__(self, job_id, func, *args, **kwargs):
         # Assign a default uuid in case of empty name
         self._job_id = job_id or uuid.uuid4()
+        # Function to be executed, define the job purpose
         self._func = func
+        # Positional arguments
         self._args = args
+        # Keywords arguments
+        self._delay = int(kwargs.pop('delay', 0))
         self._kwargs = kwargs
         # Start time of the job placeholder
         self._start_time = None
@@ -65,15 +69,35 @@ class Job:
     def end_time(self):
         return self._end_time
 
+    @property
+    def delay(self):
+        return self._delay
+
     def finished_ok(self):
         if self._status == JobStatus.OK:
             return True
         return False
 
     def execution_time(self):
-        return self._end_time - self._start_time
+        """Return the time passed between the start of the execution and the end of it"""
+        try:
+            return self._end_time - self._start_time
+        except TypeError:
+            return 0
+
+    def add_delay(self, delay):
+        """Add a delay to the function"""
+        self._delay = delay
 
     def execute(self):
+        """Execute the function with arguments and keyword arguments, if a given delay is specified,
+        this method await till the timeout expires and then executes the job"""
+        if self._delay:
+            time.sleep(self._delay)
+            return self._execute_job()
+        return self._execute_job()
+
+    def _execute_job(self):
         """Execute the function with arguments and keyword arguments"""
         self._start_time = time.time()
         try: