+++ tasq/settings.py
@@ -8,7 +8,7 @@ tasq.settings.py
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import os
-from . import Configuration
+from .util import Configuration
 
 
 class TasqConfig(Configuration):